refactor : PlayerGroundPoint 뎁스 제한을 위해 위젯으로 분리했습니다
