+++ src/1_Page/Championship/ui/DashBoard/ui/MatchListTab/ui/MatchLineupContainer/ui/FootballGroundSection/index.tsx
@@ -1,5 +1,5 @@
 import { formation } from "../../../../../../../../../../4_Shared/constant/formation";
-import PlayerGroundPoint from "./ui/PlayerGroundPoint";
+import PlayerGroundPoint from "../../../../../../../../../../2_Widget/PlayerGroundPoint";
 
 const FootballGroundSection = (props: FootballGroundSectionProps) => {
   const { championshipDetail } = props;
